# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "이화여대 창업 교육 LMS API",
        "title": "EarnLearning LMS API",
        "contact": {},
        "version": "1.0"
    },
    "host": "earnlearning.com",
    "basePath": "/api",
    "paths": {
...
        "/admin/grants": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "관리자용: 새 정부과제 생성",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "정부과제 생성",
                "parameters": [
                    {
                        "description": "과제 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.CreateGrantRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/grants/{id}/approve/{appId}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "관리자용: 정부과제 지원 승인",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "정부과제 지원 승인",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "과제 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "지원 ID",
                        "name": "appId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/admin/grants/{id}/close": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "관리자용: 정부과제 종료",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "정부과제 종료",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "과제 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/grants": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "정부과제 목록 조회 (페이지네이션)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Grant"
                ],
                "summary": "정부과제 목록",
                "parameters": [
                    {
                        "type": "string",
                        "description": "상태 필터",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "페이지",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "크기",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/grants/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "정부과제 상세 정보 조회",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Grant"
                ],
                "summary": "정부과제 상세",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "과제 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
        "/grants/{id}/apply": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "정부과제에 지원",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Grant"
                ],
                "summary": "정부과제 지원",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "과제 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "지원 정보",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.ApplyGrantRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_interfaces_http_handler.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_interfaces_http_handler.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_INPUT"
                },
                "message": {
                    "type": "string",
                    "example": "잘못된 입력입니다"
                }
            }
        },
        "internal_interfaces_http_handler.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/internal_interfaces_http_handler.APIError"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "internal_interfaces_http_handler.ApplyGrantRequest": {
            "type": "object",
            "properties": {
                "proposal": {
                    "type": "string",
                    "example": "지원 제안서"
                }
            }
        },
...
        "internal_interfaces_http_handler.CreateGrantRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "지원 설명"
                },
                "max_applicants": {
                    "type": "integer",
                    "example": 5
                },
                "reward": {
                    "type": "integer",
                    "example": 100000
                },
                "title": {
                    "type": "string",
                    "example": "창업지원금"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}